requirements.txt
fastapi
uvicorn
easyocr
opencv-python-headless
pillow
numpy
python-multipart
pytest
httpx
requests
python-dotenv